feat : LineName VO 구현
